test(server): build every test server from one constructor
Three near-identical constructors had drifted apart: one set no health
checker, another enabled a different set of grants, and the sessions config
was spelled out in each. Collapse them onto a shared builder that takes the
connectors to serve, leaving three thin wrappers for what actually differs.

Give the test sessions a real cookie encryption key. The forged-cookie tests
were reading the key back off the server they were exercising, which the
Server no longer carries: the session config was a field only the wiring
used.
