+++ server/server_logout_test.go
@@ -82,7 +82,7 @@ func TestHandleLogoutNoHintPOSTPerformsLogout(t *testing.T) {
 	req := httptest.NewRequest("POST", "/logout", nil)
 	req.AddCookie(&http.Cookie{
 		Name:  "dex_session",
-		Value: internal.SessionCookieValue(userID, connectorID, nonce, server.sessionConfig.CookieEncryptionKey),
+		Value: internal.SessionCookieValue(userID, connectorID, nonce, testSessionKey),
 	})
 	server.ServeHTTP(rr, req)
 
@@ -372,7 +372,7 @@ func TestHandleLogoutFromCookie(t *testing.T) {
 	req := httptest.NewRequest("POST", "/logout", nil)
 	req.AddCookie(&http.Cookie{
 		Name:  "dex_session",
-		Value: internal.SessionCookieValue(userID, connectorID, nonce, server.sessionConfig.CookieEncryptionKey),
+		Value: internal.SessionCookieValue(userID, connectorID, nonce, testSessionKey),
 	})
 	server.ServeHTTP(rr, req)
 
@@ -402,7 +402,7 @@ func TestLogoutCallbackWithExpiredSession(t *testing.T) {
 	req := httptest.NewRequest("GET", "/logout/callback", nil)
 	req.AddCookie(&http.Cookie{
 		Name:  "dex_session",
-		Value: internal.SessionCookieValue("user-1", "mock", "nonce", server.sessionConfig.CookieEncryptionKey),
+		Value: internal.SessionCookieValue("user-1", "mock", "nonce", testSessionKey),
 	})
 	server.ServeHTTP(rr, req)
 	require.Equal(t, http.StatusBadRequest, rr.Code)
