+++ server/server_test.go
@@ -32,11 +32,9 @@ import (
 
 	"github.com/dexidp/dex/connector"
 	"github.com/dexidp/dex/connector/mock"
-	"github.com/dexidp/dex/pkg/featureflags"
 	"github.com/dexidp/dex/server/connectors"
 	"github.com/dexidp/dex/server/device"
 	"github.com/dexidp/dex/server/oauth2"
-	"github.com/dexidp/dex/server/session"
 	"github.com/dexidp/dex/server/signer"
 	"github.com/dexidp/dex/server/tokens"
 	"github.com/dexidp/dex/storage"
@@ -82,138 +80,6 @@ E2qpAoGAZo5Wwwk7q8b1n9n/ACh4LpE+QgbFdlJxlfFLJCKstl37atzS8UewOSZj
 FDWV28nTP9sqbtsmU8Tem2jzMvZ7C/Q0AuDoKELFUpux8shm8wfIhyaPnXUGZoAZ
 Np4vUwMSYV5mopESLWOg3loBxKyLGFtgGKVCjGiQvy6zISQ4fQo=
 -----END RSA PRIVATE KEY-----`)
-
-func newTestServer(t *testing.T, updateConfig func(c *Config)) (*httptest.Server, *Server) {
-	var server *Server
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		server.ServeHTTP(w, r)
-	}))
-
-	logger := newLogger(t)
-	ctx := t.Context()
-
-	sig, err := signer.NewMockSigner(testKey)
-	if err != nil {
-		t.Fatalf("failed to create mock signer: %v", err)
-	}
-
-	config := Config{
-		Issuer:  s.URL,
-		Storage: memory.New(logger),
-		Web: WebConfig{
-			Dir: "../web",
-		},
-		Logger:             logger,
-		PrometheusRegistry: prometheus.NewRegistry(),
-		HealthChecker:      gosundheit.New(),
-		SkipApprovalScreen: true, // Don't prompt for approval, just immediately redirect with code.
-		AllowedGrantTypes: []string{ // all implemented types
-			oauth2.GrantTypeDeviceCode,
-			oauth2.GrantTypeAuthorizationCode,
-			oauth2.GrantTypeClientCredentials,
-			oauth2.GrantTypeRefreshToken,
-			oauth2.GrantTypeTokenExchange,
-			oauth2.GrantTypeImplicit,
-			oauth2.GrantTypePassword,
-		},
-		Signer: sig,
-	}
-	if updateConfig != nil {
-		updateConfig(&config)
-	}
-	s.URL = config.Issuer
-
-	// Default rotation policy, set before the server is built so the token
-	// endpoint captures it.
-	if config.RefreshTokenPolicy == nil {
-		config.RefreshTokenPolicy = tokens.NewRefreshStrategy(true, 0, 0, 0, config.Now)
-	}
-
-	// Mirror cmd: the session config is present iff the sessions feature flag is on.
-	if featureflags.SessionsEnabled.Enabled() && config.SessionConfig == nil {
-		config.SessionConfig = &session.Config{CookieName: "dex_session", AbsoluteLifetime: 24 * time.Hour, ValidIfNotUsedFor: time.Hour}
-	}
-
-	connector := storage.Connector{
-		ID:              "mock",
-		Type:            "mockCallback",
-		Name:            "Mock",
-		ResourceVersion: "1",
-	}
-	if err := config.Storage.CreateConnector(ctx, connector); err != nil {
-		t.Fatalf("create connector: %v", err)
-	}
-
-	if server, err = newServer(ctx, config); err != nil {
-		t.Fatal(err)
-	}
-
-	return s, server
-}
-
-func newTestServerMultipleConnectors(t *testing.T, updateConfig func(c *Config)) (*httptest.Server, *Server) {
-	var server *Server
-	s := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		server.ServeHTTP(w, r)
-	}))
-
-	logger := newLogger(t)
-	ctx := t.Context()
-
-	sig, err := signer.NewMockSigner(testKey)
-	if err != nil {
-		t.Fatalf("failed to create mock signer: %v", err)
-	}
-
-	config := Config{
-		Issuer:  s.URL,
-		Storage: memory.New(logger),
-		Web: WebConfig{
-			Dir: "../web",
-		},
-		Logger:             logger,
-		PrometheusRegistry: prometheus.NewRegistry(),
-		Signer:             sig,
-		SkipApprovalScreen: true, // Don't prompt for approval, just immediately redirect with code.
-	}
-	if updateConfig != nil {
-		updateConfig(&config)
-	}
-	s.URL = config.Issuer
-
-	if config.RefreshTokenPolicy == nil {
-		config.RefreshTokenPolicy = tokens.NewRefreshStrategy(true, 0, 0, 0, config.Now)
-	}
-
-	// Mirror cmd: the session config is present iff the sessions feature flag is on.
-	if featureflags.SessionsEnabled.Enabled() && config.SessionConfig == nil {
-		config.SessionConfig = &session.Config{CookieName: "dex_session", AbsoluteLifetime: 24 * time.Hour, ValidIfNotUsedFor: time.Hour}
-	}
-
-	connector := storage.Connector{
-		ID:              "mock",
-		Type:            "mockCallback",
-		Name:            "Mock",
-		ResourceVersion: "1",
-	}
-	connector2 := storage.Connector{
-		ID:              "mock2",
-		Type:            "mockCallback",
-		Name:            "Mock",
-		ResourceVersion: "1",
-	}
-	if err := config.Storage.CreateConnector(ctx, connector); err != nil {
-		t.Fatalf("create connector: %v", err)
-	}
-	if err := config.Storage.CreateConnector(ctx, connector2); err != nil {
-		t.Fatalf("create connector: %v", err)
-	}
-
-	if server, err = newServer(ctx, config); err != nil {
-		t.Fatal(err)
-	}
-	return s, server
-}
 
 func TestNewTestServer(t *testing.T) {
 	newTestServer(t, nil)
